+++ solvers/ADRSolver/EquationSystems/SteadyAdvectionDiffusionReaction.h
@@ -31,6 +31,7 @@
 // Description: Steady advection-diffusion-reaction solve routines
 //
 ///////////////////////////////////////////////////////////////////////////////
+
 #ifndef NEKTAR_SOLVERS_ADRSOLVER_EQUATIONSYSTEMS_ADVECTIONDIFFUSIONREACTION_H
 #define NEKTAR_SOLVERS_ADRSOLVER_EQUATIONSYSTEMS_ADVECTIONDIFFUSIONREACTION_H