+++ CHANGELOG.md
@@ -49,6 +49,7 @@ v5.5.0
 - Refactor UnsteadyReactionDiffusion as a subclass of UnsteadyDiffusion (!1713)
 - Tidy-up ADRSolver (!1715)
 - Refactor UnsteadyAdvectionDiffusion as a subclass of UnsteadyAdvection (!1720)
+- Refactor UnsteadyViscousBurgers as a subclass of UnsteadyInviscidBurgers solver (!1718)
 
 **CardiacEPSolver**
 - Fix cell model history point filter output after base class change (!1342)